requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
astunparse==1.6.3
Authlib==1.5.2
boto3==1.37.38
botocore==1.37.38
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cleantext==1.1.4
click==8.1.7
cmake==4.0.0
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
dateparser==1.2.0
distlib==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.8
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.25.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.3
nltk==3.9.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pipenv==2024.4.0
platformdirs==4.3.6
protobuf==4.25.5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.10
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
s3transfer==0.11.5
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.6
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.27.1
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
